+++ metadata.rb
@@ -4,7 +4,7 @@ license          "Apache 2.0"
 description      "installs/configures dop_base"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 name             "dop_base"
-version          "0.1.1"
+version          "0.1.2"
 
 %w{ debian }.each do |os|
 	supports os
@@ -22,6 +22,7 @@ end
 	htpasswd
   chef_handler
   chef-sugar
+  sysctl
 	}.each do |ressource|
 	depends ressource
 end